fixed unit tests to succeed on a Mac PPC machine (which is big-endian rather than little-endian)
git-svn-id: https://svn.urbansim.org/trunk@11457 1513fa5d-83ce-422e-ae47-6967bb0fb594

+++ opus_core/store/flt_storage.py
@@ -212,7 +212,7 @@ class StorageTests(opus_unittest.OpusTestCase):
         
     def test_load_table(self):
         expected = {
-            'city_id': array([3, 1, 2], dtype=int32),
+            'city_id': array([3, 1, 2], dtype='<i4'),
             'city_name': array(['Unknown', 'Eugene', 'Springfield']),
             }
         actual = self.storage.load_table('cities')
